Suggestion on demographic line
If we're sticking to the "use plain language" tenet of Gov.uk's styleguide I think "demographic" is too marketing of a term for "types of humans". To me it feels less inclusive simply because it's a bit of a jargon word. Being really explicit about which humans you're talking about and then leaving this as an open doc means there's room to add and evolve. Thanks for the offer to participate!