Add a Spanish version of the Code of Conduct
I've added a Spanish version of the Code of Conduct along with some supporting changes to make the repo simpler to navigate.

- Added code-of-conduct-es.md: this is the main Spanish translation
- Added code-of-conduct-en.md: this is the new home for the English version of the CoC
- Refactored the README.md to help navigate to the language versions of the CoC

Why refactor the README? The main reason is that long READMEs often go unread and may be intimidatting to new users. It also helps to be friendly right of the bat. The different links point to the language version but also give pointers on how people can contribute to them.